[Fix] use cropper, and clean server.py
